refactor(core): extract pure keyboard navigation into utils/
Moves resolveVizelListNavigation and resolveVizelGridNavigation from
controllers/listboxController.ts into utils/keyboard-navigation.ts.
These functions are pure resolvers with no DOM access, so they belong
in the utils/ layer per Section 1 of the v2.0.0 design spec.

Leaves controllers/listboxController.ts as a thin re-export shim.
Section 3 of the redesign replaces this file with the actual
createVizelListboxController factory.

Updates packages/core/src/index.ts to expose the helpers through the
Utils export block instead of the Controllers block.
